Fix bug in sensitivity analysis of C1122
The second term in the numerator of the expression for the sensitivity of C1122 contained E*(1-v) as a coefficient for Dv terms, when it should have been E*v. This has no effect for the SIMP model for which Dv is zero, but it effects the final sensitivity Dc of the compliance, and therefore the update computed for each design, including the initial design. Comparison of results before and after the fix show that this change is relatively minor; but it reduces the iteration count for some models significantly since the fixed-point iteration scheme converges more quickly.