Fix implementation of lattice stiffness models
The models for the Young's and shear moduli of the isotruss, octet truss, and ORC truss are both normalized by the appropriate modulus of the lattice constituent material; that is, the models are of $E^h/E_S$ and $G^h/G_S$, respectively. This was incorrectly implemented as E = Es * term1 + term2 + ... + termN, when it should have been implemented as E = Es * (term1 + term2 + ... + termN) where the sum of terms represents the model. Depending on the magnitudes of Es and Gs, this represents a significant error. In the case of the paper by Watts et al (see README), Es = 1.0 and Gs = 0.34, so that the value of E was correct, but G was too high by a factor of about 3.