Correct Jacobian determinant
For brevity, the integration of the finite element stiffness matrix is largely hard-coded. The determinant of the Jacobian of the isoparametric mapping was hard-coded by dividing the rest of the integrand by 4, which is appropriate for 2D elements of uniform size 1x1, rather than the correct value of 8 for 3D elements with uniform size 1x1x1. This change fixes the error.